+++ src/usecase/tag.rs
@@ -1,2 +1,2 @@
+pub mod list;
 pub mod lookup_backlinks;
-pub mod usecase;
